backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
playwright>=1.40.0
python-dotenv>=1.0.0
